Made path work with deep hierarchies. Added PML for IllegalClockCrossings
